Update the e2e test for the premium theme with-theme flow
This PR adds a quick workaround for the e2e test to avoid failing tests after the new theme tier implementation is released to users.

The implementation changes the UI for the badges which is the UI we used in this test to determine a premium theme that will be used in the flow.