Two Factor Security Keys: Automatically prompt for a security key.
The current 2fa authentication flow with security keys require users to
click on the "Continue with security key" button before we attempting to
initiate this flow.

This change automatically asks for a security key if the given account
supports it. If it fails for some reason, the old button will still work
as always if the user decides to retry the same action.