Allow showing intro offer notice for non-jetpack products
The previous hasIntroductoryDiscount was misleadingly named and located
in a file that didn't really need it. This way it is closer to where it
is used and is more clear about what it does.

Previously it only returned true for Jetpack products, but now it allows
all products with intro offers. Previously it would also show "discount
for first term" even if the intro offer applied to multiple terms, but
now it will only show if the offer is for one term.