Signup: Add presales to DIFM (Built By) flows
Implements peCdcN-lE-p2

Opted to add a new flag at the level of flow declarations. This way we
avoid a lot of duplicated code. The possible downside would be that we
cannot enable it per step. But we don't have a use case for that, nor
are we planning for it (either the whole flow has presales or not).