E2E: Fix Jetpack Related Posts block test (#87388)
In Automattic/jetpack#35353 the "display heading" setting was removed in
favor of including a heading sub-block that users can configure
directly. Fill that now.

Note this does the minimum needed to get the existing test to pass. If
someone wants to create a test for a block without a heading, they
should add that in a separate PR.