E2E: Fix Jetpack Related Posts block test again (#87465)
The fix in #87388 was accidentally looking for a related post block
inside of another related post block (because `context.addedBlockLocator`
already points at the just-added block, then `.getByRole()` was looking
inside of that). That normally wouldn't work, but due to a bug in
Automattic/jetpack#35353, that's exactly the DOM structure that existed
in the page. 😀

Now that we're fixing that bug with Automattic/jetpack#35686, the test
needs fixing too.