Remove one-off theme purchases from customizer
One-off purchases of individual themes were removed last year so that
themes were only available by buying the relevant plan or themes bundle.
It is still possible though to make the customizer display a purchase
link for individual themes.

This change (when deployed after a server side counterpart) prevents
that, it sends the user to buy the lowest necessary WordPress.com plan
for the theme they're looking at.