Updates site migration instructions for key retrieval
When the user first lands on the page, there is a "Get migration key" button. Clicking the button retrieves the key from the API and displays it (hidden) along with "Show key" and "Copy key" buttons.

In the event that there's an API error (either because of a lack of permissions or the key has already been retrieved), we show instructions and a link for manually retrieving the key instead.