Avoid reader shortcuts conflicting with cmd palette (#88348)
cmd+k is used to launch the wpcom command palette. `k` is also used to
move between posts in the reader. This meant that toggling the command
palette also changed the post being viewed in the background.

This change avoids that by disabling the reader keyboard shortcuts when
the ctrl or cmd keys are being held.