Reduce size of images in designSetup by passing smaller arguments to mShot.
This is a follow-up to @niranjan-uma-shankar finding that the design
picker has a 70 MB payload. We can see in our funnel data that this
causes a lot of users, including new users, to drop the process of
creating a site at this step. When segmenting users, we can see that
users are relatively more likely to drop out if they have a slow
connection or a slow device, which supports the idea that 70 MB is too
large a payload for many users.

This commit reduces the payload size to 16 MB, which is still too large
a payload for many users, but will allow more users to be able to pick
a theme and create a site. There are a number of possible longer-term
solutions that could get this payload down to the point where almost
every user will be able to pick a theme, but those will be more involved
fixes.