E2E: Fix broken use of `for...in` rather than `for...of`
The `for...in` loop used in this test iterates over the keys of the
array, not the values. The test happens to pass as long as the post
contains "0" and "1" (e.g. in the date or title) and fails otherwise.

What we really want to do is iterate over the values from the `split`.