E2E: Fix click location in FullSiteEditorDataViewsComponent
Prior to WordPress/gutenberg#59803, the `a` inside the
`.dataviews-view-table__primary-field` div occupied the full width of
the div, so when Playwright clicked in the center of the div it would
hit the link.

That PR changed things up a bit, now the `a` is not expanded to fill the
entire div and a click in the middle of the div misses the link, meaning
the test never navigates to the next page as expected.

This updates the logic to click the actual `a` instead of the div.