Create valid checkout links for theme related plan upgrades (#89544)
Users of the DesignPicker with a > 1 year plan could not use a theme from a
higher tier plan (except for partner themes). The design picker would send
them to checkout with a 1 year plan and the checkout doesn't allow changes to
a shorter plan term.

This change fixes that problem by sending users to checkout with an upgraded
plan with the same term as their current plan. It also displays the details
of that plan on the theme upgrade modal, rather than it displaying the
details for the annual plan.

It does this by using the `minimumUpsellPlan` from the theme's tier and then
finding a plan in the plan list of the equivalent length. This replaces the
existing hard-coding, as well as the partner plan calculation.

The calculated plan's slug is then passed to the <ThemeUpgradeModal> using
the `requiredPlan` prop, which replaces the earlier `currentPlanTerm` prop.
The ThemeUpgradeModal then uses this to display the correct price and refund
information (which is still hard-coded inside the ThemeUpgradeModal).