Update site-migration-assign-trial-plan step to allow direct-linking
Previously, the step only tried to assign the trial plan once when the step first renders. This was fine for progressing through the flow normally but doesn't work if we try to directly link to the step.

Now we check the `site` dependency so we can try to assign the trial plan again after any updates.