Create valid checkout links for plan upgrades in lits (#89798)
Users of the Logged-in theme showcase with a > 1 year plan could not use a
theme from a higher tier plan. They'd be sent to the checkout but the
checkout doesn't allow upgrades to plans with shorter terms.

This change fixes that problem by sending users to checkout with an upgraded
plan with the same term as their current plan.