Optionally skip polling if the email is verified (#89817)
The import-verify-email step, by default, will poll continuously to see if the email has been verified. Once the email is verified, the `submit` function is triggered.

This allows us to optionall disable polling. In the case of the site-migration flow, we don't want to submit the step since the user is going to be redirect back to the flow after clicking on the verification email.