E2E: Update test to still pass with released versions of Gutenberg
PR #90225 updated a test to pass in the run against the nightly
Gutenberg release after WordPress/gutenberg#60632 changed some labels.
Unfortunately that broke the runs against released versions of
Gutenberg.

This adjusts the test to match both the old and new labels, with a TODO
comment pointing out that it can be adjusted once that Gutenberg change
is released to all the environments we test against.