Hide site with unknown capabilities from domain-only flow
During the domain-only flow we show a list of existing sites the user
can attach their new domain to. If the site details are not loaded
correctly, and the `capabilities` field is missing, then we should treat
this as if the user doesn't have the required permission to manage the
site and hide it from the list.

I don't know how this situation happens in practice, but Sentry
identified that it's possible at runtime.