Reduce number of domains placeholders (#90759)
While waiting for domains info to load, placeholders for eleven domains
were displayed. If the person only had 1-2 domains this caused a small
amount of jumping around once they were loaded.

Having 1-2 domains is normal for a given site, and seems like a
reasonable default for the global domains list.