Fix lack of "/setup" prefix when code is running on the browser.
Due to some weird issue, when the location.pathname runs on the browser context, it doesnt return the react router base path, in our case /setup.
The best way if found to fix it is using the react-router useHref to generate the path.

I added some tests to cover this case because is very easy to break a lot of stuff and very hard to debug.