Add an environment dropdown in the site dashboard
Instead of displaying a separate site entry for staging sites, we now display group sites with their corresponding staging site in a single entry. This is done by adding a dropdown with the site environment containing two options: Production and Staging.

When the user selects the staging option, they will be sent to the staging site detail page.