Command palette: Add support for wildcard keywords  (#91453)
Commands can now opt-in to being treated as wildcards which lets them
better match search terms which might be used. This is particularly
useful for where we want to map wp-cli commands: searching for 
`wp user delete` will now match the manage users command whereas 
previously it would stop matching after `wp user `.

Commands can opt-in to the wildcard search by ending part of a 
searchLabel with an asterisk.