Move `recordCommandPaletteOpen` to a separate file
Moving `recordCommandPaletteOpen` to a separate file avoids build
failures. Probably the failures are because the build process isn't
expecting to be processing this CSS at this time, just for access to
that function.

The command palette's index.tsx should be refactored in due course, but
for now, this seems to work.