Rebrand Quick actions to command palette  (#91948)
This will introduce the idea of the command palette to people so that
they can explore it. The Quick Actions panel is renamed and details on
how to open the command palette are added.

Several command aliases have also been added so that searching for
listed quick actions match a command.