DIFM Design Picker: Don't use incorrect constants (#92125)
ca95596 changed the way that tracks properties were calculated. To do
this it used two constants - PREMIUM_THEME and MARKETPLACE_THEME however
these constants are related to theme types rather than tiers.

It hasn't caused a problem in practice - PREMIUM_THEME resolves to
'premium' (the tier name) so that worked, and though MARKETPLACe_THEME
resolves to 'marketplace' when the tier is called 'partner', that didn't
matter because there are no marketplace themes in the DIFM design
picker.