Disable refs to jetpackedgephp74 and jetpackedge82 pending rename (#92327)
We're planning to rename these sites to jetpackedgephpold and
jetpackedgephpnew to better reflect their purpose. To avoid potential
disruption, though, we have to comment out the references to the old
names while we're doing that.

Here we don't bother to remove the refs from parameter validation and
configurations, just the places where it's actually used by CI.

See p1720019588866209-slack-C05Q5HSS013 for details.