e2e: Fix Jetpack Forms test (#92538)
The test using AI to create the form seems to have been broken since
Automattic/jetpack#37754 and Automattic/jetpack#38067, which switched
the AI Assistant from a custom form extension to a generic one. This
changed the placeholder the test was looking for, and also the "busy"
behavior is different.

TeamCity didn't notice this test failing, though. It turns out that's
because the test for creating a form using patterns and the test for
creating a form using AI are both named like
`Jetpack Forms (Atomic: default): Add and configure blocks in the editor: Form: Configure the block`
so TeamCity decided that the two separate tests were really just one
test being flaky, and that it passed as long as one of the 'copies'
passed (probably because the same thing is used to handle automatic
retries when a test really is being flaky).

To fix that, we add a new optional `blockTestName` field that can be
used to explicitly name the tests instead of using the name of the
block. They're now
`Jetpack Forms (Atomic: default): Add and configure blocks in the editor: Form (Patterns): Configure the block`
and
`Jetpack Forms (Atomic: default): Add and configure blocks in the editor: Form (AI): Configure the block`

And for good measure, add some code to try to detect if this happens
again.