e2e: Fix instagram and twitter block flows tests (#92749)
It appears that #92736 updated the tests for WordPress/gutenberg#63052
to make the Gutenberg edge tests pass, and in doing so broke tests
against Atomic and Simple that haven't been updated to that version yet.

Update the tests to temporarily accept either option. This can be
reverted once Atomic and Simple are updated to (AFAICT) Gutenberg
v18.8.0.