Stats: Update paywall upgrade notice (#93003)
* Force the paywall upgrade notice

* Update title for error state

* Update body for error state

The notice is now explicit about the features moving behind the paywall and surfaces the date in the message text.

* Add emphasis to paywall date

* Remove notice override

* Localize displayed paywall date

* Use upgrade_deadline_date value from API

No longer need to calculate an end date on the client side as the API provides this now.

* Update comments, variable name

* Guard against empty value from API

The value is defined as optional so we need to guard against that. Now we only show the date in the notice body if we have one.