Partner themes: Default plan lengths to a year
Partner themes require a subscription to the theme (can be monthly or
yearly) and an atomic capable subscription (monthly, yearly, 2yearly,
3yearly etc). In the theme showcase partner themes were defaulting to
having an annual theme subscription and a monthly plan subscription.

This is now changed so that they both default to annually:
 * We prefer subscribers to have longer term plans
 * Cancelling the monthly sub would give you an annual theme sub that
   you couldn't use.

Note that these are default values: if you already have a paid plan
then a plan of a matching length will be added.