All flows using first_post_published path returned by server
The first_post_published task was using the calypso_path returned by the
server, expect for `starting-writing` and `design-first` flows, which
wanted to always open the wp-admin editor.
In Automattic/jetpack#39259 the server now always returns a wp-admin
path for these flows, so the special casing is no longer required.