Fix rendering of null values in performance profiler history
Null-value rendering was changed in #94773 so that nulls appeared as
gaps in the graph. This change was only made to the "V1" version of the
components i.e. the logged-out view.

As of #95108 the logged-out view uses the "V2" version of components and
so this fix was lost in the process.

This diff applies the fix from #94773 to the "V2" component.