e2e: Fix Map block test
Automattic/jetpack#39708 associated the "Add a location" `<label>` with
the `<input>`, which makes Playwright use that as the accessible name
for the input rather than the input's placeholder.

Switch the test to look for the placeholder specifically, instead of the
name, which will work with both the old and new code.