e2e: Further fix flaky Twitter test
PR #95579 attempted to adjust the test to accept both the expected
iframe and the fallback link produced when Twitter's API is flaky.
Unfortunately the way Playwright interpreted the code is like
`iframe > ( tweet or link )` rather than `( iframe > tweet ) or link` as
intended, so it's still not finding the link.

Fall back to using `Promise.any()` to combine the two locators instead,
like we do in a bunch of other places.