Update step navigation controls
The "Next step" button is now in the header. The "Skip to dashboard"
button is at the bottom of the page.

Although the mockup calls the skip button "Go to WordPress Admin", we
aren't yet ready to default all new sites to wp-admin, so I'll keep
calling it "Dashboard" for now.