Ensure new sell goals select the sell flow (for now)
The sell flow will go away, however for now we want to ensure selecting
a sell goals sets the sell intent, so that proceeding in the flow goes
to the sell step.

I was able to tidy the code. There might have been a point in time when
the GOAL_TO_INTENT_MAP was useful, however now all it was doing was
setting the write intent.