Replace `createStore` import with `legacy_createStore as createStore`
According to docs (https://redux.js.org/usage/migrations/migrating-rtk-2):
```
createStore Deprecation
In Redux 4.2.0, we marked the original createStore method as @deprecated. Strictly speaking, this is not a breaking change, nor is it new in 5.0, but we're documenting it here for completeness.

This deprecation is solely a visual indicator that is meant to encourage users to migrate their apps from legacy Redux patterns to use the modern Redux Toolkit APIs.

The deprecation results in a visual strikethrough when imported and used, like createStore, but with no runtime errors or warnings.

createStore will continue to work indefinitely, and will not ever be removed. But, today we want all Redux users to be using Redux Toolkit for all of their Redux logic.

To fix this, there are three options:

Follow our strong suggestion to switch over to Redux Toolkit and configureStore
Do nothing. It's just a visual strikethrough, and it doesn't affect how your code behaves. Ignore it.
Switch to using the legacy_createStore API that is now exported, which is the exact same function but with no @deprecated tag. The simplest option is to do an aliased import rename, like import { legacy_createStore as createStore } from 'redux'
```