Remove 'prorated-introductory-offer' from codes which signify offers
The `prorated-introductory-offer` cost override code is a prorated
discount for the time remaining within the period of time covered by the
introductory offer pricing, not an introductory offer discount itself.
The difference is subtle but important and we should not treat it as an
introductory offer for all the possible places that
`isOverrideCodeIntroductoryOffer()` might be used.