Move loading PayPalScriptProvider to PayPalProvider
We have to do this in a new wrapper file between `CheckoutMainWrapper`
and `CheckoutMain` because it needs access to the shopping cart provided
by `CheckoutMainWrapper` but also needs to be present for
`useCreatePaymentMethods` which happens inside `CheckoutMain`.

We should probably move all payment method wrappers (eg:
`StripeHookProvider`) into this file but we'll start with just this one.