Marketplace Theme Install: Fix stuck loading screen
When a user adds a Partner theme during onboarding (Design Picker) and then removes the Partner theme from the checkout and purchases only the Business plan, the user is redirected to /marketplace/thank-you page.

Since the theme was removed, the page loads indefinitely for the Atomic transfer and the theme installation to finish - thus, blocking the user on this page.