Hide "Remove domain" option for 100-year domains that have auto-renew turned off
100-year domains should never have auto-renew turned off, and D165092-code
added checks on the backend to prevent that. Turning it off would enable
the "Remove domain" option in the purchase settings page.

If, for some reason, auto-renew is turned off for a 100-year domain, this
PR hides the "Remove domain" button.

This PR also replaces the "Domain expires on" label by "Paid until" for
100-year domains that are going to expire in 3 months or less.