Site settings: introduce new selector to detect wpcom
Fixes https://github.com/Automattic/at-help/issues/897

Before this commit, we relied on the isJetpackSite selector to detect whether we should display the Holiday Snow settings or not (we want to display them on Wordress.com sites only (simple OR Atomic).

This commit introduces a new site selector that will return true when we're interested in a WordPress.com Site (simple OR atomic), and starts using it.

Epic: https://github.com/Automattic/dotcom-forge/issues/10082