Improve consistency of the login form
Previously, if the user had used the magic-link before, we should them a "Previously used: Email me a magic link" button and changed the button on the right from "Email me a login link" to "Continue with email".

This has led to confusion with some users being unsure of how to login with a password.

To that end, we've changed it so that the "Email Address or Username" is visible when the previously-used login method was magic-login. The social button on the right will also always read "Email me a login link".

We're not moving the "Password" input at this point since there's a lot of infrastructure for other login flows that depends on the password being entered in a second step.

All other social logins will remain the same and appear under "Previously used".