Add/big sky before plans (#99019)
* Add big sky to goals-first onboarding (behind feature flag)

* Existing design choice screen uses site-setup-flow.ts to navigate to Big Sky

The Big Sky selection on the design choice screen started bypassing the
flow logic in this commit:
https://github.com/Automattic/wp-calypso/commit/ba9641060527779cb7200e99680e06816d975eb5#diff-a9c1ce51ac0fbf560d78a2bbbb650b6edd25cb528c840a1e435dfb752c2eae4fR131

This moves the logic back into site-setup-flow.ts. This allows us to
have different logic when the design choice step is used in the
site-setup flow vs. the onboarding flow.

* Enable Big Sky in onboarding in development and wpcalypso

* Save design choice selection to onboarding store

It would have been more generic to store an enum of design choice
selections in the store. However I think "design choice" would get
confused with "selected design".

* Send goal-first users to Big Sky if that is their selection

* Removed unnecessary logic

* Persisting createWithBigSky so onboarding will remember that config

---------

Co-authored-by: Paulo Trentin <paulo@paulotrentin.com.br>