Fix "Use your domain" component being displayed briefly after leaving flow
This is caused because we're assuming `leaveFlowFunction()` is going to be called and then we'd be leaving the flow. However, since the function runs the function to update the state right after that, sometimes the component can be visible for a short period (one second or so) and only then the flow gets exited. This commit fixes this by halting the function flow with an early return.