Remove newsletter goal feature flag (#99445)
This change removes the feature flag for the newsletter goal from the
onboarding flow. The flag is no longer necessary as the feature we are
enabling the feature for all users.

I left the ability to mock feature flags in the `goalsToIntent()`. Seems
like it could be useful in the future.