Launchpad requests goals-first cumulative experience
This change refactors the `launchpadContext` code so it can be aware
of whether the user should see cumulative improvements to the
goals-first flow.

The new hook `useGoalsFirstCumulativeExperience`
is introduced to determine if the user should see the cumulative
improvements. This replaces the use of `useGoalsFirstExperiment()` for
determining whether the user is in the cumulative group or not. The new
hook also supports working outside of the stepper framework, since some
cumulative improvements will be outside of the stepper.