Actions: add a new workflow to enforce use of Type labels on PRs
In our Quality reports, we want to be able to know what percentage of our PRs are bug fixes vs. enhancements vs. new features.
We can rely on the existing [Type] labels for that, but for reporting to be accurate we want each and every PR to use such a label from now on.

This workflow will fail whenever such a label is not found on the PR. We can make the workflow required for merges to trunk once this has been merged.

Of note, this requirement is also enforced in the Jetpack repo (although via the Repo Gardening action instead of a custom workflow):

- https://github.com/Automattic/jetpack/pull/40428
- https://github.com/Automattic/jetpack/pull/42260