Fix: Loss of playgroundId GET param during login screens in playground onboarding flow
If the user is logged out, the user may lose GET param because of the possible combinations of navigation between diff user screens
where the user can possibly login from diff methods available, create an account, recover an account, change their intented action after choosing it once etc.

To ensure we never lose it, despite new user login screens that may be added in future,
we are storing playgroundId in localstorage on playground step and consume it on domains step.

The entire flow still relies on playgroundId GET param in URL, including the domains step.
But now localstorage can be used to recover the lost value, if its lost.