Stale bot: update general rules.
Internal reference: p7H4VZ-5is-p2

**For issues**

Right now in the Calypso repo, we leave a comment on issues and mark them as stale after 180 days. pdqkMK-1rQ-p2
Our dotcom PAs then triage those issues and close them if necessary. Let's instead make it clear to the issue creator that it's on them to update the issue, and add a rule to close those issues a month later.

**For PRs**

Right now in the Calypso repo, we don't do anything for PRs.

Let's instead leave a comment after a month, and then close the PR a month later. Folks will have a full month to come back to their PR and update it if necessary.