PoweredBy: Normalise logo sizes
For the landscape aspect ratio logos, Use the size prop as the height
and calculate the width from that. This makes it easier to have
consistent vertical sizes of logos, while horizontally they can take the
space they need.

Needs further testing to see if any existing uses of the logos now
display incorrectly.