Allow user cancellation of DIFM migrations
This adds a button to the site overview in the case where the user has a
DIFM migration started.

Clicking the button displays a dialog which lets the user send a request
to cancel the migration. This calls the
/automated-migration/cancel-migration endpoint.

We display a success or error notice depending on the result from the
endpoint.

If the request is successful, the user will see the normal site overview
on refresh.