Remove Button and style overrides from "Visit stats" link
The "Visit stats" link was centre aligned when it should have been left
aligned. This is because the Button uses inline-flex.

We were already overriding default Button styles to get the link look we
wanted, so better to use a plain <a> and get the left-alignment we want
for free.