Backup: move 'isFinished' to the state
When doing a granular restore, we cannot rely on the
'inProgressRewindStatus' value alone to decide whether to show the
finished/success page, given it might be storing stale information from
a previous run. Ensuring the 'isFinished' value is reset on every
restore and only using 'inProgressRewindStatus' to change its value at
the actual execution end is a safer approach.