Update URL keys to be unique to the DataView
This prevents calypso from re-using the keys from one DataView in
another when navigating between two (eg: from Billing History to Active
Upgrades).

This also will allow us to have the URL preserve the view for multiple
DataViews on the same page if we ever get around to implementing it for
the Memberships table which is on the same page as the purchases table.