Revert "Hosting Dashboard: Link plan card to billing section (#105670)"
This works fine in the v2 hosting dashboard. It doesn't work fine in the
v1 dashboard though.

The v2 hosting dashboard has been retrofitted to the v1 sites overview,
but the billing section it links to is limited to the v2 hosting
dashboard and so the link does nothing in the v1 sites overview.

This reverts commit a52f52101effd21761ca75070d36720a1062d57d.