Hosting Dashboard: Adding performance page WIP
Much of the advanced functionality is placeholder at the moment.
I was unable to find a satisfactory search component to use, and I'm not
sure that the ComboboxControl component is flexible to meet the design.
I hoped to use the Tabs WP component to implement the selection tabs
left of the graph card, but it is not available yet, so an alternative
will be needed. The FoldingCard component is also unavailable. I wasn't
sure if there is an existing good solution to use for the collapsing
recommendations. If needed, I can just manually code up a toggle.