Limit plans in 'plan-upgrade' flow depending upon feature
Setting a feature= GET var when accessing the plan-upgrade flow, will
limit the plans displayed further to just the users plan + the plans
that provide the feature that are upgrades to the users current plan.

E.g. if supplying the `sftp-and-database-access` feature then only
display the current plan + atomic-capable plans.