Remove premium suggestion from suggestions list if user searched for a premium FQDN
Sometimes we get suggestions from our domain suggestion providers that are
premium but not marked as such. This is a problem in onboarding, for example,
where we don't offer premium domains. It was confusing for users because if
they search for a FQDN, we run an availability check and show an error notice
if the domain is premium. However, that domain remained in the results list
as available.