Move logic from useAssertConditions to initialize
useAssertConditions was already deprecated, moving code to initalize is
the right thing to do.

This has required changing types - initialize was checking for false
before, but it wasn't a possible return value, now it is.

The login check has been removed because stepsWithRequiredLogin already
takes care of it.

Also standardised on url query var rather than param.