Fix error in type_check_client
The window object can not be extended because, in the editor, it
believes that the `currentUser` property has already been set (by a type
definition in `data-stores`). However if we don't extend the type, then
the type_check_client script doesn't see the `currentUser` property
definition and it errors.

This should be fixed by aligning all the various user types together.