Prevent NS records from pointing to *.wordpress.com
NS records can't point to WPCOM name servers since that doesn't make
sense - if the DNS of a domain is managed here, that means it's already
using our name servers.

The DNS record form for NS records showed "Invalid host" if a user
inputted any host ending in `.wordpress.com`, but it didn't prevent the
record from being added. That's because the validation wasn't in the
correct place.