Increase clarity around retired themes
On the theme details screen don't show a theme badge for retired themes
unless it is the users active theme. This reduces confusion where users
may think that upgrading plan will give them access to the retired
theme, while continuing to be clear that where it is the active theme,
that it's use may be contingent on their current plan.