Update the theme component for retired themes
Pass the new `isThemeRetired` and `isThemeActiveForSite` props to
ThemeTierBadge from the Theme component.

This is a 'belt and braces' change, as this component is currently only
used in theme grids, which only display retired themes if they're
active, in which case the active badge takes precedence over the badge
here anyway.