Improve alignment of retirement notice
The WordPress Notice component assumes it should be the full width of
the container. This doesn't look correct for the /themes page.

An earlier commit added alignment for the global logged-in showcase but
the logged-out showcase was left full width and the with-site pages were
not exactly lined up.

This change brings the same consist alignment from the global logged-in
showcase to all of the other versions of the theme showcase.